refactor: rename to less specific
